requirements.txt
# Core Data Engineering
# Note: Python 3.13 has limited PySpark support. Consider using Python 3.11 or 3.12 for best compatibility.
pyspark==3.5.8
delta-spark==3.2.1
py4j==0.10.9.7
pandas==2.1.0
numpy==1.24.3

# Data Generation
faker==19.3.1

# AWS Integration (optional)
boto3==1.28.25
botocore==1.31.25

# Data Quality
great-expectations==0.18.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Code Quality
black==23.7.0
flake8==6.1.0

# Notebook Support
jupyter==1.0.0
ipython==8.14.0

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
